reuse common edit components
