+++ admin-ui/src/component/reservations/queries.ts
@@ -1,19 +1,5 @@
 import { gql } from "@apollo/client";
-import {
-  RESERVATION_META_FRAGMENT,
-  RESERVATION_UNIT_FRAGMENT,
-} from "./fragments";
-
-export const RESERVATION_COMMON_FRAGMENT = gql`
-  fragment ReservationCommon on ReservationType {
-    pk
-    createdAt
-    state
-    begin
-    end
-    orderStatus
-  }
-`;
+import { RESERVATION_COMMON_FRAGMENT } from "./fragments";
 
 export const RESERVATIONS_QUERY = gql`
   ${RESERVATION_COMMON_FRAGMENT}
@@ -70,30 +56,6 @@ export const RESERVATIONS_QUERY = gql`
         hasNextPage
       }
       totalCount
-    }
-  }
-`;
-
-// TODO do we need user / orderStatus?
-export const SINGLE_RESERVATION_QUERY = gql`
-  ${RESERVATION_META_FRAGMENT}
-  ${RESERVATION_UNIT_FRAGMENT}
-  ${RESERVATION_COMMON_FRAGMENT}
-  query reservationByPk($pk: Int!) {
-    reservationByPk(pk: $pk) {
-      ...ReservationCommon
-      type
-      workingMemo
-      reservationUnits {
-        ...ReservationUnit
-      }
-      user {
-        firstName
-        lastName
-        email
-        pk
-      }
-      ...ReservationMetaFields
     }
   }
 `;
